// src/types/auth.ts
import { DefaultSession } from 'next-auth';

interface AuthUser {
  name: string;
  email: string;
  picture?: string | null;
  image?: string | null;
  accessToken: string;
  // refreshToken: string;
  sub: string;
  expires_at: number;
}

export interface AuthSession extends Omit<DefaultSession, 'user'> {
  user: AuthUser;
}